Fix trailing newlines in JSON, add format test against ref
Pretty print on block and quote close was adding newlines because it was
transitioning to StateMain which considered it as first whitespace and
printed anyway. However, whitespaces after end of block or string is
unnecessary. This was fixed by adding transitioning from *Close states to
the StateWhitespace that eats up whitespaces.

FIXME: This fixed JSON but broke HCL.

Also, added test that compares formatting result with reference. To make
it work, Runner now accepts io.Writer output instead of always printing
to stdout.